Copy the implementation from the use-spoon2 branch
Import spoon in a backwards compatible manner, so that if spoon is not
installed then we just use the previous server code.

Also remove the forked server, since this is a simple mix-in that a
subclass can add, and it keeps the backwards compatibility code simple.
It’s likely that the subclass wants to choose how many workers anyway.